Fix issue with string builder removing too much raw sln data on Unix platforms
The implementations incorrectly assume there are two trailing characters to ignore for the newline (i.e. "\r\n"), which is the case for Windows, however this is only a single character on Unix platorms ("\n").
This fix uses Environment.NewLine.Length to determine how many trailing characters to ignore, depending on which platform is being used.